Adjust trigger for deploying to preview
Under active development there may be multiple PRs being merged in
rapid succession and each of these will trigger a deploy currently
and there may be failures as only one deploy will work at a time.

By switching to deploying on release we can control when the deploy
happens and batch up a group of changes to deployed at the same time.

The trigger is when the release is 'created' with the idea that we
will create draft releases that deploy to preview automatically and
then when we have a production environment we will publish the
release to deploy to production.