Make factories (mostly) consistent
In a number of places we are doing hacky things to ensure petitions
are in a consistent state for the tests we're doing. This commit removes
most of those aside from instances where we're checking things like
signature counting jobs, invalidation, etc.

This also picked up a few places where the test setup was incorrect.