feat: let admins copy the content of a GB petition to GD
If the Gaelic website is disabled, then only the English interface is
available. However members of the public might still submit a petition
in Gaelic, which means an English-version of the petition is created
with Gaelic content in it. To help deal with the mismatch, offer a
"Copy content to Gaelic petition" button that just does that.