Prefer sass-ruby over sassc-ruby
The latter takes 4 minutes to build on the EC2 instance the first time
it is booted which means 4 minutes longer for it to be ready in the
event of a scaling requirement. The speed up over the pure Ruby version
is negligible for the quantity of assets we have so revert back.