Add the option to disable threshold and debate features
Previously we've set the threshold timestamps to the same values as
open_at/closed_at but this commit changes that since we're now removing
them from the JSON/CSV data and skipping them in `increment_count!`.
The only other option would've been to increase the threshold for a
debate to a ludicrous number which didn't seem right.