Resend the correct email to petition creators
When we're not collecting signatures the correct email to resend is the
validate signature email and not the gather sponsors email.

The rate limiting inside the email job has been removed as this is
handled within the petition creator form directly.

We also don't send a copy to the feedback address as moderators can
validate petition creators signatures directly - the copy of the
gather sponsors email is sent so that moderators could have access
to the 'Sponsor this petition' link.