petition: ensure :all facet has a deterministic order
we now order the :all petitions facet by most recently published which
orders petitions on their `open_at` timestamp. This is fine on the
frontend (`open_at` is required to appear there) but in the admin some
petitions might not have the attribute set which results in an
non-deterministic order. Add an extra criteria to resolve this (and
reassure the tests).