Ensure that template names are on the defined list
Some of the templates have been deleted in the live list but may need
to be brought back at some point. To assist in this change the name
field to a select which has all the templates known to the system and
disable the ones that have already been created.