Allow capitalised local parts in email addresses
Validating the imported emails raised some issues so allow uppercased
local parts which is allowed. Most mail servers are case insensitive
so we previously downcased to prevent people abusing case to create
multiple email addresses but adding a functional index and the addition
of the canonical email feature means that it isn't an issue anymore.