Don't encode HTML entities in JSON by default
The Rails default is to encode HTML entities so that calling to_json in
a view with inline <script> tags is secure by default. However, we only
use that pattern in a single place so manually escape it there and disable
the default encoding so that URLs in API results don't encode '&'.