Prevent double count query
The `empty?` method of the will_paginate gem calls `count` as does the
`total_entries`. The latter is cached whereas the former isn't so rather
than delegating we can rewrite the method to call `total_entries` which
prevents a double count query on the petitions index page. It won't
really affect performance since the query was in the Active Record
cache but it should stop Appsignal warning us about an N+1 query.