Use a separate session-like cookie for signed tokens
We need to use SameSite=Strict for the session cookie settings so that
credentials used by moderators are protected but this breaks our method
for detecting when someone is sharing the signed url. Fix this by moving
signed_tokens into its own session-like cookie which we can set to 'Lax'
and will be available after clicking a link in a verification email.