Skip redirect to new sponsor page when no sponsors required
There's a 15 minute window on the verify link normally before a petition
is published so that people don't share their verify links and get confused
when their signatures are added. However, when sponsors aren't required
this shouldn't be applied as the creator could be redirected to the new
sponsor page if they click the link again in their email.

Also block any requests to the sponsors controller when not collecting sponsors.